+++ code/FESL/server/commands/basic/handle_add_sub_account.cpp
@@ -22,7 +22,7 @@ namespace FESL {
 		if(kv_list.HasKey("TID")) {
 			tid = kv_list.GetValueInt("TID");
 		}
-		request.extra = (void *)tid;
+		request.extra = (void *)(ptrdiff_t)tid;
 		request.peer = this;
 		request.peer->IncRef();
 		request.type = TaskShared::EProfileSearch_CreateProfile;
@@ -38,9 +38,9 @@ namespace FESL {
 			((Peer *)peer)->mp_mutex->lock();
 			((Peer *)peer)->m_profiles.push_back(results.front());
 			((Peer *)peer)->mp_mutex->unlock();
-			((Peer *)peer)->SendError(FESL_TYPE_ACCOUNT, FESL_ERROR_NO_ERROR, "AddSubAccount", (int)extra);
+			((Peer *)peer)->SendError(FESL_TYPE_ACCOUNT, FESL_ERROR_NO_ERROR, "AddSubAccount", (int)(ptrdiff_t)extra);
 		} else {
-			((Peer *)peer)->handle_web_error(error_details, FESL_TYPE_ACCOUNT, "AddSubAccount", (int)extra);
+			((Peer *)peer)->handle_web_error(error_details, FESL_TYPE_ACCOUNT, "AddSubAccount", (int)(ptrdiff_t)extra);
 		}
 	}
 }
